chore: move notify_user_job to another file
